+++ example/lib/pages/welcome_page.dart
@@ -8,6 +8,7 @@ import 'package:better_player_example/pages/change_player_theme_page.dart';
 import 'package:better_player_example/pages/controller_controls_page.dart';
 import 'package:better_player_example/pages/controls_always_visible_page.dart';
 import 'package:better_player_example/pages/controls_configuration_page.dart';
+import 'package:better_player_example/pages/dash_page.dart';
 import 'package:better_player_example/pages/drm_page.dart';
 import 'package:better_player_example/pages/event_listener_page.dart';
 import 'package:better_player_example/pages/fade_placeholder_page.dart';
@@ -155,6 +156,9 @@ class _WelcomePageState extends State<WelcomePage> {
       }),
       _buildExampleElementWidget("DRM", () {
         _navigateToPage(DrmPage());
+      }),
+      _buildExampleElementWidget("DASH", () {
+        _navigateToPage(DashPage());
       }),
     ];
   }