+++ lib/src/asms/better_player_asms_audio_track.dart
@@ -18,5 +18,11 @@ class BetterPlayerAsmsAudioTrack {
   ///mimeType of the audio track
   final String? mimeType;
 
-  BetterPlayerAsmsAudioTrack({this.id, this.segmentAlignment, this.label, this.language, this.url, this.mimeType});
+  BetterPlayerAsmsAudioTrack(
+      {this.id,
+      this.segmentAlignment,
+      this.label,
+      this.language,
+      this.url,
+      this.mimeType});
 }