+++ lib/src/asms/better_player_asms_data_holder.dart
@@ -1,6 +1,6 @@
-import 'better_player_asms_track.dart';
-import 'better_player_asms_audio_track.dart';
-import 'better_player_asms_subtitle.dart';
+import 'package:better_player/src/asms/better_player_asms_audio_track.dart';
+import 'package:better_player/src/asms/better_player_asms_subtitle.dart';
+import 'package:better_player/src/asms/better_player_asms_track.dart';
 
 class BetterPlayerAsmsDataHolder {
   List<BetterPlayerAsmsTrack>? tracks;