+++ lib/src/asms/better_player_asms_track.dart
@@ -21,7 +21,8 @@ class BetterPlayerAsmsTrack {
   ///mimeType of the video track
   final String? mimeType;
 
-  BetterPlayerAsmsTrack(this.id, this.width, this.height, this.bitrate, this.frameRate, this.codecs, this.mimeType);
+  BetterPlayerAsmsTrack(this.id, this.width, this.height, this.bitrate,
+      this.frameRate, this.codecs, this.mimeType);
 
   factory BetterPlayerAsmsTrack.defaultTrack() {
     return BetterPlayerAsmsTrack('', 0, 0, 0, 0, '', '');