+++ lib/src/dash/better_player_dash_utils.dart
@@ -1,7 +1,3 @@
-// Dart imports:
-import 'dart:convert';
-import 'dart:io';
-
 // External Package imports:
 import 'package:better_player/src/asms/better_player_asms_data_holder.dart';
 import 'package:better_player/src/core/better_player_utils.dart';
@@ -15,18 +11,18 @@ import 'package:better_player/src/asms/better_player_asms_track.dart';
 
 ///DASH helper class
 class BetterPlayerDashUtils {
-
   static Future<BetterPlayerAsmsDataHolder> parse(
       String data, String masterPlaylistUrl) async {
     List<BetterPlayerAsmsTrack> tracks = [];
-    List<BetterPlayerAsmsAudioTrack> audios = [];
-    List<BetterPlayerAsmsSubtitle> subtitles = [];
+    final List<BetterPlayerAsmsAudioTrack> audios = [];
+    final List<BetterPlayerAsmsSubtitle> subtitles = [];
     try {
       int audiosCount = 0;
       final document = XmlDocument.parse(data);
       final adaptationSets = document.findAllElements('AdaptationSet');
       adaptationSets.forEach((node) {
         final mimeType = node.getAttribute('mimeType');
+
         if (mimeType != null) {
           if (MimeTypes.isVideo(mimeType)) {
             tracks = tracks + parseVideo(node);
@@ -34,76 +30,90 @@ class BetterPlayerDashUtils {
             audios.add(parseAudio(node, audiosCount));
             audiosCount += 1;
           } else if (MimeTypes.isText(mimeType)) {
-            subtitles.add(parseSubtitle(node));
+            subtitles.add(parseSubtitle(masterPlaylistUrl, node));
           }
         }
       });
     } catch (exception) {
-      BetterPlayerUtils.log("Exception on hls parse: $exception");
+      BetterPlayerUtils.log("Exception on dash parse: $exception");
     }
-    return BetterPlayerAsmsDataHolder(tracks: tracks, audios: audios, subtitles: subtitles);
+    return BetterPlayerAsmsDataHolder(
+        tracks: tracks, audios: audios, subtitles: subtitles);
   }
 
   static List<BetterPlayerAsmsTrack> parseVideo(XmlElement node) {
-    List<BetterPlayerAsmsTrack> tracks = [];
+    final List<BetterPlayerAsmsTrack> tracks = [];
 
     final representations = node.findAllElements('Representation');
 
     representations.forEach((representation) {
       final String? id = representation.getAttribute('id');
       final int width = int.parse(representation.getAttribute('width') ?? '0');
-      final int height = int.parse(representation.getAttribute('height') ?? '0');
-      final int bitrate = int.parse(representation.getAttribute('bandwidth') ?? '0');
-      final int frameRate = int.parse(representation.getAttribute('frameRate') ?? '0');
+      final int height =
+          int.parse(representation.getAttribute('height') ?? '0');
+      final int bitrate =
+          int.parse(representation.getAttribute('bandwidth') ?? '0');
+      final int frameRate =
+          int.parse(representation.getAttribute('frameRate') ?? '0');
       final String? codecs = representation.getAttribute('codecs');
       final String? mimeType = MimeTypes.getMediaMimeType(codecs ?? '');
-      tracks.add(BetterPlayerAsmsTrack(id, width, height, bitrate, frameRate, codecs, mimeType));
+      tracks.add(BetterPlayerAsmsTrack(
+          id, width, height, bitrate, frameRate, codecs, mimeType));
     });
 
     return tracks;
   }
 
   static BetterPlayerAsmsAudioTrack parseAudio(XmlElement node, int index) {
-    String segmentAlignmentStr = node.getAttribute('segmentAlignment') ?? '';
+    final String segmentAlignmentStr =
+        node.getAttribute('segmentAlignment') ?? '';
     String? label = node.getAttribute('label');
-    String? language = node.getAttribute('lang');
-    String? mimeType = node.getAttribute('mimeType');
+    final String? language = node.getAttribute('lang');
+    final String? mimeType = node.getAttribute('mimeType');
 
-    if (label == null) {
-      label = language;
-    }
+    label ??= language;
 
     return BetterPlayerAsmsAudioTrack(
-      id: index,
-      segmentAlignment: segmentAlignmentStr.toLowerCase() == 'true',
-      label: label,
-      language: language,
-      mimeType: mimeType
-    );
+        id: index,
+        segmentAlignment: segmentAlignmentStr.toLowerCase() == 'true',
+        label: label,
+        language: language,
+        mimeType: mimeType);
   }
 
-  static BetterPlayerAsmsSubtitle parseSubtitle(XmlElement node) {
-    String segmentAlignmentStr = node.getAttribute('segmentAlignment') ?? '';
+  static BetterPlayerAsmsSubtitle parseSubtitle(
+      String masterPlaylistUrl, XmlElement node) {
+    final String segmentAlignmentStr =
+        node.getAttribute('segmentAlignment') ?? '';
     String? name = node.getAttribute('label');
-    String? language = node.getAttribute('lang');
-    String? mimeType = node.getAttribute('mimeType');
-    String? url = node.getElement('Representation')?.getElement('BaseURL')?.text;
-    if (url != null && url.startsWith('//')) {
-      url = 'https:' + url;
+    final String? language = node.getAttribute('lang');
+    final String? mimeType = node.getAttribute('mimeType');
+    String? url =
+        node.getElement('Representation')?.getElement('BaseURL')?.text;
+    if (url?.contains("http") == false) {
+      final Uri masterPlaylistUri = Uri.parse(masterPlaylistUrl);
+      final pathSegments = <String>[...masterPlaylistUri.pathSegments];
+      pathSegments[pathSegments.length - 1] = url!;
+      url = Uri(
+              scheme: masterPlaylistUri.scheme,
+              host: masterPlaylistUri.host,
+              port: masterPlaylistUri.port,
+              pathSegments: pathSegments)
+          .toString();
     }
 
-    if (name == null) {
-      name = language;
+    if (url != null && url.startsWith('//')) {
+      url = 'https:$url';
     }
+
+    name ??= language;
 
     return BetterPlayerAsmsSubtitle(
-      name: name,
-      language: language,
-      mimeType: mimeType,
-      segmentAlignment: segmentAlignmentStr.toLowerCase() == 'true',
-      url: url,
-      realUrls: [url ?? '']
-    );
+        name: name,
+        language: language,
+        mimeType: mimeType,
+        segmentAlignment: segmentAlignmentStr.toLowerCase() == 'true',
+        url: url,
+        realUrls: [url ?? '']);
   }
-
 }